Refactoring queries.
